Dropping support for < 1.13.0
