+++ .travis.yml
@@ -12,9 +12,7 @@ env:
     - EMBER_TRY_SCENARIO=ember-release
     - EMBER_TRY_SCENARIO=ember-beta
     - EMBER_TRY_SCENARIO=ember-canary
-    - EMBER_TRY_SCENARIO=ember-1.10.0
-    - EMBER_TRY_SCENARIO=ember-1.11.3
-    - EMBER_TRY_SCENARIO=ember-1.12.0
+    - EMBER_TRY_SCENARIO=ember-1.13.0
 
 matrix:
   fast_finish: true
