--- addon/computeds/moment.js
@@ -1,6 +1,0 @@
-import Ember from 'ember';
-import momentFormat from './format';
-
-Ember.deprecate('ember-moment: `ember-moment/computeds/moment` is deprecated in favor of `ember-moment/computeds/format`');
-
-export default momentFormat;
