+++ addon/utils/compute-fn.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
 
-export default function(cb, globalAllowEmpty = false) {
+export default function(cb) {
   return function(params, hash) {
     if (!params || params && params.length === 0) {
       throw new TypeError('ember-moment: Invalid Number of arguments, expected at least 1');
@@ -11,7 +11,7 @@ export default function(cb, globalAllowEmpty = false) {
     let allowEmpty = hash.allowEmpty || hash['allow-empty'];
 
     if(allowEmpty === undefined || allowEmpty === null) {
-      allowEmpty = !!globalAllowEmpty;
+      allowEmpty = !!this.get('globalAllowEmpty');
     }
 
     if (allowEmpty && [null, '', undefined].indexOf(datetime) > -1) {
