--- app/helpers/duration.js
@@ -1,4 +1,0 @@
-import durationHelper from 'ember-moment/helpers/deprecated/duration';
-import makeBoundHelper from 'ember-moment/utils/make-bound-helper';
-
-export default makeBoundHelper(durationHelper);
