+++ app/helpers/moment-duration.js
@@ -1,4 +1,1 @@
-import durationHelper from 'ember-moment/helpers/moment-duration';
-import makeBoundHelper from 'ember-moment/utils/make-bound-helper';
-
-export default makeBoundHelper(durationHelper);
+export { default } from 'ember-moment/helpers/moment-duration';
