+++ app/helpers/moment-format.js
@@ -1,9 +1,8 @@
 import Ember from 'ember';
-import momentFormatHelper from 'ember-moment/helpers/moment-format';
-import makeBoundHelper from 'ember-moment/utils/make-bound-helper';
 import config from '../config/environment';
+import Helper from 'ember-moment/helpers/moment-format';
 
-const { get } = Ember;
-const helper = momentFormatHelper(get(config, 'moment.outputFormat'), !!get(config, 'moment.allowEmpty'));
-
-export default makeBoundHelper(helper);
+export default Helper.extend({
+  globalOutputFormat: Ember.get(config, 'moment.outputFormat'),
+  globalAllowEmpty: !!Ember.get(config, 'moment.allowEmpty')
+});
