--- app/modern/helpers/moment-duration.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-moment/helpers/moment-duration';
