--- app/modern/helpers/moment-format.js
@@ -1,8 +1,0 @@
-import Ember from 'ember';
-import config from '../config/environment';
-import Helper from 'ember-moment/helpers/moment-format';
-
-export default Helper.extend({
-  globalOutputFormat: Ember.get(config, 'moment.outputFormat'),
-  globalAllowEmpty: !!Ember.get(config, 'moment.allowEmpty')
-});
