--- app/modern/helpers/moment-to-now.js
@@ -1,7 +1,0 @@
-import Ember from 'ember';
-import config from '../config/environment';
-import Helper from 'ember-moment/helpers/moment-to-now';
-
-export default Helper.extend({
-  globalAllowEmpty: !!Ember.get(config, 'moment.allowEmpty')
-});
