+++ config/ember-try.js
@@ -1,23 +1,9 @@
 module.exports = {
   scenarios: [
     {
-      name: 'ember-1.10.0',
+      name: 'ember-1.13.0',
       dependencies: {
-        'ember': '1.10.0',
-        'ember-load-initializers': 'ember-cli/ember-load-initializers#0.0.2'
-      }
-    },
-    {
-      name: 'ember-1.11.3',
-      dependencies: {
-        'ember': '1.11.3',
-        'ember-load-initializers': 'ember-cli/ember-load-initializers#0.0.2'
-      }
-    },
-    {
-      name: 'ember-1.12.0',
-      dependencies: {
-        'ember': '1.12.0'
+        'ember': '1.13.0'
       }
     },
     {
