+++ tests/dummy/app/templates/index.hbs
@@ -1,43 +1,25 @@
-{{moment now}}
-
-<code>
-  \{{moment-format now}}
-</code>
-
-<hr>
-
-{{moment-format now 'LLLL' }}
-
+{{moment-format now 'LLLL'}}
 <code>
   \{{moment-format now 'LLLL'}}
 </code>
 
 <hr>
 
-{{moment-format usIndependenceDay 'LLLL' }}
-
+{{moment-format usIndependenceDay 'LLLL'}}
 <code>
   \{{moment-format usIndependenceDay 'LLLL'}}
 </code>
 
 <hr>
 
-{{moment-format usIndependenceDay 'MMM DD, YYYY' }}
-
+{{moment-format usIndependenceDay 'MMM DD, YYYY'}}
 <code>
   \{{moment-format usIndependenceDay 'MMM DD, YYYY'}}
 </code>
 
 <hr>
 
-{{moment-format emptyDate allow-empty=true}}
-{{moment-to-now emptyDate allow-empty=true}}
-{{moment emptyDate allow-empty=true}}
-{{ago emptyDate allow-empty=true}}
-{{moment-from-now emptyDate allow-empty=true}}
-
 {{moment-format now 'LL' 'LLLL'}}
-
 <code>
   \{{moment-format now 'LL' 'LLLL'}}
 </code>
@@ -45,7 +27,6 @@
 <hr>
 
 {{moment-from-now now}}
-
 <code>
   \{{moment-from-now now}}
 </code>
@@ -53,7 +34,6 @@
 <hr>
 
 {{moment-from-now lastHour interval=1000}}
-
 <code>
   \{{moment-from-now lastHour interval=1000}}
 </code>
@@ -61,7 +41,6 @@
 <hr>
 
 {{moment-from-now usIndependenceDay}}
-
 <code>
   \{{moment-from-now usIndependenceDay}}
 </code>
